refactor(citation): relocate CitationInstance to the citation app
Move CitationInstance (with its manager, slug minting and validators)
from provenance down into citation, beside CitationSource -- the
citation app now owns the whole evidence domain. With the claim FK
gone, nothing in the model reaches up: the one place that must know
both Claim and CitationInstance is the ClaimCitationInstance join,
which stays in provenance and imports citation downward. Import-linter
confirms all 26 contracts, so the spine is now honest.

The relocation is state-only: SeparateDatabaseAndState migrations adopt
the model in citation (over the existing, unrenamed
provenance_citationinstance table) and release it from provenance
(retargeting the join FK and Claim's citation_instances M2M). The one
real database operation renames the django_content_type row in place
(provenance -> citation): a ContentType's identity is (app_label,
model), and minting a fresh row instead would dangle every inline
wikilink's RecordReference.target_type and split the "what links here"
graph. Verified live: a pre-migration reference row keeps its pk across
the rename and its footnote still renders.

The cite LinkType and picker registrations move to the citation app
config and retarget to citation.CitationInstance. The model's admin
moves to citation/admin.py; ReadOnlyAdminMixin moves to core so both
apps can share it. ~25 imports retarget; the model's test file moves to
citation/tests with the Claim-side M2M test staying in provenance.
