+++ backend/apps/catalog/tests/test_patch_inline_cites.py
@@ -16,13 +16,14 @@ from django.core.exceptions import ValidationError
 
 from apps.catalog.models import Manufacturer
 from apps.catalog.tests.conftest import make_machine_model
+from apps.citation.models import CitationInstance
 from apps.claim_ingest.apply import apply_plan
 from apps.claim_ingest.patches import PatchError, build_plan, load_patch
 from apps.core.markdown import (
     convert_storage_to_authoring,
     render_all_links,
 )
-from apps.provenance.models import CitationInstance, Source
+from apps.provenance.models import Source
 
 pytestmark = pytest.mark.django_db
