+++ backend/apps/core/admin.py
@@ -1,6 +1,25 @@
 from django.contrib import admin
+from django.db.models import Model
+from django.http import HttpRequest
 
 from .models import License
+
+
+class ReadOnlyAdminMixin:
+    """Disallow add / change / delete so the admin is view-only."""
+
+    def has_add_permission(self, request: HttpRequest) -> bool:
+        return False
+
+    def has_change_permission(
+        self, request: HttpRequest, obj: Model | None = None
+    ) -> bool:
+        return False
+
+    def has_delete_permission(
+        self, request: HttpRequest, obj: Model | None = None
+    ) -> bool:
+        return False
 
 
 @admin.register(License)
