+++ backend/apps/provenance/admin.py
@@ -1,33 +1,15 @@
 from django.contrib import admin
-from django.db.models import Model
-from django.http import HttpRequest
+
+from apps.core.admin import ReadOnlyAdminMixin
 
 from .models import (
     ChangeSet,
-    CitationInstance,
     Claim,
     ClaimCitationInstance,
     IngestRun,
     Source,
     SourceFieldLicense,
 )
-
-
-class ReadOnlyAdminMixin:
-    """Disallow add / change / delete so the admin is view-only."""
-
-    def has_add_permission(self, request: HttpRequest) -> bool:
-        return False
-
-    def has_change_permission(
-        self, request: HttpRequest, obj: Model | None = None
-    ) -> bool:
-        return False
-
-    def has_delete_permission(
-        self, request: HttpRequest, obj: Model | None = None
-    ) -> bool:
-        return False
 
 
 class SourceFieldLicenseInline(admin.TabularInline[SourceFieldLicense, Source]):
@@ -110,25 +92,6 @@ class ClaimAdmin(ReadOnlyAdminMixin, admin.ModelAdmin[Claim]):
         if len(s) > 80:
             return s[:80] + "..."
         return s
-
-
-@admin.register(CitationInstance)
-class CitationInstanceAdmin(ReadOnlyAdminMixin, admin.ModelAdmin[CitationInstance]):
-    """Read-only inspection view. CitationInstances are immutable.
-
-    An instance is shared evidence with no single owning claim; the claims
-    citing it are inspected through the ClaimCitationInstance admin.
-    """
-
-    list_display = ("citation_source", "slug", "locator_truncated", "created_at")
-    list_select_related = ("citation_source",)
-    readonly_fields = ("citation_source", "slug", "locator", "created_at")
-
-    @admin.display(description="Locator")
-    def locator_truncated(self, obj: CitationInstance) -> str:
-        if len(obj.locator) > 60:
-            return obj.locator[:60] + "..."
-        return obj.locator
 
 
 @admin.register(ClaimCitationInstance)
