+++ backend/apps/provenance/models/__init__.py
@@ -10,7 +10,6 @@ from ..model_bases import (
     LinkableLifecycleClaimModel,
 )
 from .changeset import CHANGESET_NOTE_MAX_LENGTH, ChangeSet, ChangeSetAction
-from .citation_instance import CITATION_INSTANCE_LOCATOR_MAX_LENGTH, CitationInstance
 from .claim import (
     Claim,
     ExistingClaimRow,
